+++ db/migrate/20260122144603_add_descriptions_to_services.rb
@@ -1,0 +1,5 @@
+class AddDescriptionsToServices < ActiveRecord::Migration[8.1]
+  def change
+    add_column :services, :description, :text
+  end
+end