+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update GOV.UK Frontend and CCS Frontend ([PR 2251](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2251))
+- Add description attribute to services ([PR 2257](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2257))
 
 ## [6.5.0] - 2025-12-09